I made the switch
I swapped the naming of the SLGetItem and SLNextItem functions as per
TA's instructions. I'd like to mention here that I feel the comments
that describe the functions gave very very bad representations of what
the functions were supposed to do.

Also. I updated the methods to match the desired actions. That is,
SLNextItem iterates first, and then returns the item in question. (Which
is stupid in my opinion)

I updated the ReadMe and published it as a PDF